Created module for emulating complex operations with real interleaved arrays
This module pulls all of the original emulation stuff from fft and
condenses things into three multiplication (overloaded) operators. All
of the sections using the emulation stuff have been updated. A basic
test was also performed and showed that the results with the new
module/operators provides the exact same results as was produced with
the original implementation.

Old TSUM related stuff was removed from Makefile