Officially removing trees RNS support from lesgo
Since the RNS stuff provided by Stuart's trees stuff was never used we
are removing it. All RNS support is provided by the RNS modules. What is
retained however is the "trees_pre_ls" program which is convenient for
generating trees, cubes, etc. for use with the level set stuff. A
makefile for trees_pre_ls was also created to help separate things.