Updated convec to use dynamically allocated arrays
For now a module is not used and an explicit interface for convec was created in convec.h. Also all non-dummy arrays are allocated on first entry of subroutine and their state saved with the 'save' attribute.

It may be advantageous to go ahead and create a module for this.