Added nproc to be read in from input file
This allows preprocessing programs (trees_pre_ls, cyl_skew_pre_ls) to be able to obtain the number of processesor to be used by lesgo so that it does not have to be compiled in. There is also the side effect that this adds a safeguard to the intented number of processors.