Changed lbc_mom from string to integer. Now 'stress free' is lbc_mom=0 and 'wall' is lbc_mom=1
Several reasons for the change: 1) 'stress free' was one of the options and now with the usage of the input file all the whitespace on the line is eaten. Therefore, it was required to change the specification to a contiguous string or change it all together. 2) ubc was already specified as an integer so it seemed to make sense to go this route. 3) Using and integer is should be faster to evaluate logicals than strings (?)