This patch optimizes the access pattern of the tavg_t structure
The purpose of this is to ensure sequential access for the tavg_t structure when performing the
averaging. In order to achieve this the order in which the subtypes are defined in the tavg type
must be followed when accessing the structure. Since some variables are only defined on the k=0
level it was necessary to add an MPI only section for the summations at k=0 which the benefits of
using jzmin and jzmax are lost.

Also updated the functions in 'stat_defs' that are called from 'tavg_finalize'

NOTE: this patch will break the ability to reuse 'tavg.out' files from previous
simulations once it is applied, since the tavg_t structure has been
changed. Just be sure to discard any old restart files after applying.