Together with previous update 1. Updates have been made to Lagrangian scale dependent model in order to handle uniform inflow cases correctly. Updates for the turbulent case will hopefully follow soon. The current version of the code does not work for turbulent inflow conditions. For periodic simulations the results of LESGO should be the same and these updates should not influence your results.
While updating this routine I noticed several broken code segments which for special cases or to generate additional output, which were not working anymore. As I am not repairing code segments that are not working and are not even used anymore I simply removed then in order to increase the readability of the code. Also several optimizations have been implemented. On very small grids it seems to speed up the entire code up to 5% when the Lagrangian scale dependent sub grid model is used. I am not sure what the speed-up on larger grids is.

2. In other sub-grid models and throughout the code I commented out DYN_TN flag and corresponding code completely. It does not seem to be fully supported now anymore, at least several cases with the Lagrangian scale dependent model are not fully supported. As this can be dangerous I simply commented this out for now. Also the flag was implemented very inconsistently leading to significant overhead throughout the code, which is not contributing to the outcome of the code (see the mentioned speedup above). 

3. Fixed the configuration of the standard version such that it compiles. The CGNS version seems to have problems, at least on my account. Tony do you know what is the problem here?