Updated io.f90 file
This submission contains an updated and cleaned up io.f90 file. It now
only contains the CGNS and BINARY output. All legacy outputs have been
removed. Also the calculations of the statistics have been updated such
that these are calculated on the natural uv or w grid and interpolation
of data is limited as much as possible. This version of the code
compiles and runs without problems. Output should be tested to make
sure nothing strange is going on.