Updated tower and nacelle routines that model the nacelle and tower for an arbitrary number of turbines instead of just one. This version is coupled to the ATM and reads out the locations from the turbines and nacelle based on the input of the ATM.
Some variables, which are normally an input parameter of the actuator
disk model, are defined as constants here, i.e.
filter_cutoff
tower_size (tower_fraction)
tower_trunc
tower_alpha
tower_theta
T_avg_dim

Note that the tower and nacelle do not have any output at the moment as
this was done with tecwwrite before and the ATM version of the code is
not compatible with this library.